add block config source
