+++ app/code/Magento/Cms/Model/Config/Source/Block.php
@@ -1,0 +1,46 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+namespace Magento\Cms\Model\Config\Source;
+
+use Magento\Cms\Model\ResourceModel\Block\CollectionFactory;
+
+/**
+ * Class Block
+ */
+class Block implements \Magento\Framework\Option\ArrayInterface
+{
+    /**
+     * @var array
+     */
+    private $options;
+
+    /**
+     * @var CollectionFactory
+     */
+    private $collectionFactory;
+
+    /**
+     * @param CollectionFactory $collectionFactory
+     */
+    public function __construct(
+        CollectionFactory $collectionFactory
+    ) {
+        $this->collectionFactory = $collectionFactory;
+    }
+
+    /**
+     * To option array
+     *
+     * @return array
+     */
+    public function toOptionArray()
+    {
+        if (!$this->options) {
+            $this->options = $this->collectionFactory->create()->toOptionIdArray();
+        }
+        return $this->options;
+    }
+}
