+++ app/code/Magento/Cms/Model/ResourceModel/AbstractCollection.php
@@ -176,4 +176,32 @@ abstract class AbstractCollection extends \Magento\Framework\Model\ResourceModel
 
         return $countSelect;
     }
+
+    /**
+     * Returns pairs identifier - title for unique identifiers
+     * and pairs identifier|entity_id - title for non-unique after first
+     *
+     * @return array
+     */
+    public function toOptionIdArray()
+    {
+        $res = [];
+        $existingIdentifiers = [];
+        foreach ($this as $item) {
+            $identifier = $item->getData('identifier');
+
+            $data['value'] = $identifier;
+            $data['label'] = $item->getData('title');
+
+            if (in_array($identifier, $existingIdentifiers)) {
+                $data['value'] .= '|' . $item->getData($this->getIdFieldName());
+            } else {
+                $existingIdentifiers[] = $identifier;
+            }
+
+            $res[] = $data;
+        }
+
+        return $res;
+    }
 }
